refactor: restructure electron folder
