+++ .prettierignore
@@ -6,3 +6,4 @@ dist
 node_modules
 
 src/redux/services/validation.test.ts
+electron/main.ts
