+++ electron/app/extensions/downloadExtensionEntry.ts
