+++ electron/app/services/templateService.ts
@@ -14,12 +14,12 @@ import {
   validateTemplatePack,
 } from '@models/template';
 
-import downloadExtension from './extensions/downloadExtension';
-import downloadExtensionEntry from './extensions/downloadExtensionEntry';
-import {createFolder, doesPathExist} from './extensions/fileSystem';
-import loadExtension from './extensions/loadExtension';
-import loadMultipleExtensions from './extensions/loadMultipleExtensions';
-import {convertExtensionsToRecord, makeExtensionDownloadData} from './utils';
+import downloadExtension from '../extensions/downloadExtension';
+import downloadExtensionEntry from '../extensions/downloadExtensionEntry';
+import {createFolder, doesPathExist} from '../extensions/fileSystem';
+import loadExtension from '../extensions/loadExtension';
+import loadMultipleExtensions from '../extensions/loadMultipleExtensions';
+import {convertExtensionsToRecord, makeExtensionDownloadData} from '../utils';
 
 const TEMPLATE_PACK_ENTRY_FILE_NAME = 'monokle-template-pack.json';
 const TEMPLATE_ENTRY_FILE_NAME = 'monokle-template.json';
