+++ electron/entitlements.mac.inherit.plist
